added login lambda for Cognito authentication
