+++ HHAN_Authentication_Function/src/.npmignore
