+++ HHAN_Authentication_Function/src/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "HHAN_Authentication_Function",
+  "version": "1.0.0",
+  "description": "HHAN_Authentication_Function",
+  "main": "app.js",
+  "author": "SAM CLI",
+  "license": "MIT",
+  "dependencies": {
+    "axios": "^0.18.0",
+    "amazon-cognito-identity-js": "^1.31.0",
+    "aws-sdk": "^2.798.0"
+  },
+  "scripts": {
+    "test": "mocha tests/unit/"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^6.1.4"
+  }
+}
