+++ HHAN_Authentication_Function/template.yaml
@@ -1,16 +1,17 @@
 AWSTemplateFormatVersion: "2010-09-09"
 Transform: AWS::Serverless-2016-10-31
 Description: >
-  HHAN_Migration_Function
+  HHAN_Authentication_Function
 
-  Sample SAM Template for HHAN_Migration_Function
+  Sample SAM Template for HHAN_Authentication_Function
 
+# More info about Globals: https://github.com/awslabs/serverless-application-model/blob/master/docs/globals.rst
 Globals:
   Function:
     Timeout: 3
 
 Resources:
-  HHANMigrationFunction:
+  HHANAuthenticationFunction:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
       CodeUri: src/
