--- HHAN_Migration_Function/samconfig.toml
@@ -1,10 +1,0 @@
-version = 0.1
-[default]
-[default.deploy]
-[default.deploy.parameters]
-stack_name = "hhan-migration-function-bucket"
-s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-ds3rdp2yhcnl"
-s3_prefix = "hhan-migration-function-bucket"
-region = "us-east-1"
-confirm_changeset = true
-capabilities = "CAPABILITY_IAM"
