+++ HHAN_Migration_Function/src/config.json
@@ -1,0 +1,5 @@
+{
+  "region": "us-east-1",
+  "CLIENT_ID": "79cbq30uthlc0nalt80t9g80i",
+  "USER_POOL_ID": "us-east-1_TzP4z3oCR"
+}
